hard to refactor into handleNoValidToken
There is a lot of duplication of work in the way I tried to refactor.
The error handling code has no way to figure out what went wrong, so it
would have to reproduce all the steps of verifyJWTToken as it was
originally written to reproduce the problems. I prevented functions from
giving detailed error messages, so that hamstrung my error handler.

It may be better to just say "getDecodedToken" and let the errors from
getDecodedToken specify what went wrong, and for now handle all of the
errors in one place.